+++ noodles-bam/src/io/reader.rs
@@ -367,7 +367,7 @@ where
     /// let region = "sq0:8-13".parse()?;
     /// let query = reader.query(&header, &index, &region)?;
     ///
-    /// for result in query {
+    /// for result in query.records() {
     ///     let record = result?;
     ///     // ...
     /// }