+++ noodles-bam/src/io/reader/query.rs
@@ -52,18 +52,6 @@ where
     }
 }
 
-impl<'r, R> IntoIterator for Query<'r, R>
-where
-    R: bgzf::io::BufRead + bgzf::io::Seek,
-{
-    type Item = io::Result<Record>;
-    type IntoIter = Records<'r, R>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.records()
-    }
-}
-
 pub(crate) fn intersects(
     record: &Record,
     reference_sequence_id: usize,
@@ -227,7 +215,7 @@ mod tests {
         let query = reader.query(&header, &index, &region)?;
 
         let actual: Vec<_> = query
-            .into_iter()
+            .records()
             .map(|result| {
                 result.and_then(|record| RecordBuf::try_from_alignment_record(&header, &record))
             })